pyproject.toml
[project]
name = "expense-calculator"
version = "0.1.0"
description = "Telegram-based expense tracker with bill scanning and visualization"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.30.0",
    "python-telegram-bot>=21.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.0.0",
    "google-generativeai>=0.8.0",
    "Pillow>=10.0.0",
    "matplotlib>=3.8.0",
    "python-dotenv>=1.0.0",
]
